Updated requirements.txt to include flask-limiter. Also, moved limiter.py to helper dir for easier access.
