+++ backend/api/routes/admin_reviews.py
@@ -4,7 +4,7 @@ from api import db,admin_reviews_ns
 from api.models.cf_models import Campaigns, Users,Comments,AdminReviews,CampaignStatus
 from api.fields.adminReviewsFields import admin_reviews_data
 from email_service.email_sender import send_email
-from ...limiter import limiter
+from ..helpers.limiter import limiter
 
 
 decision_to_status = {
