+++ backend/requirements.txt
@@ -31,3 +31,5 @@ charset-normalizer>=3.0.0
 # Production server
 gunicorn==21.2.0
 requests==2.31.0  # For health checks
+
+flask-limiter
